Review findings, add reverse mapping test, move mapping into test
