+++ src/test/groovy/org/nomin/functional/MapingToMapTest.groovy
@@ -1,17 +1,21 @@
 package org.nomin.functional
 
 import java.util.HashMap
+import org.nomin.Mapping
+import org.nomin.entity.Animal
+import org.nomin.entity.Person
 import org.nomin.core.Nomin
-import org.nomin.mappings.*
-import org.nomin.entity.*
 
 /**
  * Tests mapping to a map.
  */
 class MapingToMapTest {
+
   def nomin = new Nomin(Animal2Map)
 
   def animal = new Animal(name: '42')
+
+  def map = ['name': 42]
 
   @org.junit.Test
   void testAnimal2Map() {
@@ -19,4 +23,17 @@ class MapingToMapTest {
     assert map?.size() == 1 && map['name'] == 42
   }
 
+  @org.junit.Test
+  void testMap2Animal() {
+    Animal animal = nomin.map(map, Animal)
+    assert animal?.name == '42'
+  }
+}
+
+class Animal2Map extends Mapping {
+  protected void build() {
+      mappingFor a: java.util.Map, b: Animal
+      b.name = a['name']
+      hint a: Integer
+  }
 }
